Versão 0.8.0 com implementação das anotações da Dra. Simone, enviadas por email.

+++ IntervencaoNarrativa/Control/FTINActivityFlowController.m
@@ -161,7 +161,7 @@ NSString const * kFTINViewedActivityBaseName = @"viewed_activity_";
 
 + (NSString *)keyForViewedActivityOfType:(FTINActivityType)type
 {
-	return [kFTINViewedActivityBaseName stringByAppendingFormat:@"%u", type];
+	return [kFTINViewedActivityBaseName stringByAppendingFormat:@"%ld", (long) type];
 }
 
 #pragma mark - Data control
